+++ R/create_site.R
@@ -2,14 +2,14 @@ create_site <- function(path) {
 
   chk <- check_site_rendered(path)
 
-  if (!isTRUE(chk$site)) fs::dir_create(fs::path(path, "site"))
+  if (!isTRUE(chk$site)) fs::dir_create(path_site(path))
 
   if (!isTRUE(chk$readme)) create_site_readme(path)
 
-  if (!isTRUE(chk$built)) fs::dir_create(fs::path(path, "site", "built"))
+  if (!isTRUE(chk$built)) fs::dir_create(path_built(path))
 
   if (!isTRUE(chk$description)) {
-    fs::file_create(fs::path(path, "site", "DESCRIPTION"))
+    fs::file_create(fs::path(path_site(path), "DESCRIPTION"))
     check_git_user(path)
     create_description(path)
   }