+++ tests/testthat/test-clear_site.R
@@ -7,8 +7,8 @@ test_that("the site can be cleared", {
   withr::defer(fs::dir_delete(tmp))
   expect_false(fs::dir_exists(tmp))
   res <- create_lesson(tmp)
-  expect_warning(s <- get_schedule(tmp), "set_schedule")
-  set_schedule(tmp, s, write = TRUE)
+  expect_warning(s <- get_episodes(tmp), "set_episodes")
+  set_episodes(tmp, s, write = TRUE)
 
   # Make sure everything exists
   expect_true(check_lesson(tmp))
@@ -27,11 +27,11 @@ test_that("the site can be cleared", {
   expect_setequal(built_site, fs::path(tmp, "site", c(expected, "docs")))
 
   rds <- fs::path(tmp, "site", "built", "assets", "data", "test.rds")
-  expect_length(fs::dir_ls(fs::path(tmp, "site", "built")), 2L)
+  expect_length(fs::dir_ls(fs::path(tmp, "site", "built")), 5L)
   expect_true(fs::file_exists(rds))
   expect_equal(readRDS(rds), expected)
 
-  clear_site(tmp)
+  reset_site(tmp)
 
   expect_length(fs::dir_ls(fs::path(tmp, "site")), 4L)
   expect_length(fs::dir_ls(fs::path(tmp, "site", "built")), 0L)