+++ tests/testthat/test-format_syllabus.R
@@ -8,8 +8,8 @@ test_that("the formatted syllabus renders markdown", {
   withr::defer(fs::dir_delete(tmp))
   expect_false(fs::dir_exists(tmp))
   res <- create_lesson(tmp, open = FALSE, rstudio = FALSE)
-  expect_warning(s <- get_schedule(tmp), "set_schedule")
-  set_schedule(tmp, s, write = TRUE)
+  expect_warning(s <- get_episodes(tmp), "set_episodes")
+  set_episodes(tmp, s, write = TRUE)
 
   res <- get_syllabus(tmp, questions = TRUE)
   fmt <- format_syllabus(res)